More work making the new algorithm stable on lower S/N data.
-- Use the sigma_source size for the aperture for deconvolving measurements
   rather than sigma_obs.

-- Fixed a minor bug in the Ellipse::shiftBy function.  Also, added a test
   of this function to the test suite.

-- Added ability to get pixels in an elliptical aperture, rather than
   a circular aperture.  Use this for the final shear pass.

-- Add feature of increasing fPsf by steps if initial shear measurement
   fails or produces a flag.

Also, unrelated to the above, I switched the (x,y) columns read in from the
catalog from X_IMAGE, Y_IMAGE to XMODEL_IMAGE, YMODEL_IMAGE.  These are
expected to be more accurate, and also correspond to the ALPHAMODEL_J2000,
DELTAMODEL_J2000 that we are using for (ra,dec), so now the WCS test should
have an rms of around 1.e-4 or less.


git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@6318 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/CoaddCatalog.cpp
@@ -195,8 +195,8 @@ void CoaddCatalog::readFits(const std::string& file)
 
     dbg<<"  "<<raCol<<"  "<<declCol<<std::endl;
     _skyPos.resize(nRows);
-    std::vector<float> ra;
-    std::vector<float> decl;
+    std::vector<double> ra;
+    std::vector<double> decl;
     table.column(raCol).read(ra, start, end);
     table.column(declCol).read(decl, start, end);
